Add Apache packages to mistral-api container
Some projects that use Kolla docker containers (TripleO) run
mistral-api under httpd. This patch adds the httpd package so that it
exists in the container and can optionally by used via Docker
entrypoints.

As many of these projects do not require the use of Apache (it is
opt-in) at this point not all deployment frameworks support the use of
it so the existing configuration defaults have been left as-is for now.

Change-Id: I07ffd0b398e20065fff27c30d0842b94090ca0ad
Related-Bug: #1724607
Partially-Implements: blueprint apache-packages-for-apis