Fix the Ubuntu Fluentd image
For ubuntu, we actually use fluentd 0.14.22 and ruby 2.1.10 now.
Thus, we should use the specific gem version for the plugins:

 - fluent-plugin-rewrite-tag-filter:2.0.0
According to the README.md of this plugin[1]
It shows:
----------------------------------------------------------
| fluent-plugin-rewrite-tag-filter | Fluentd    | Ruby   |
|----------------------------------|------------|--------|
| >= 2.0.0                         | >= v0.14.2 | >= 2.1 |
| < 2.0.0                          | >= v0.12.0 | >= 1.9 |
----------------------------------------------------------

 - fluent-plugin-grok-parser:2.1.4
When using the fluent-plugin-rewrite-tag-filter 2.0.0 to fix this bug, another
bug[1] which is the same as this one[2] was triggered.
According to https://github.com/fluent/fluent-plugin-grok-parser, it shows:
---------------------------------------------------
| fluent-plugin-grok-parser | fluentd    | Ruby   |
|---------------------------|------------|--------|
| >= 1.0.0                  | >= v0.14.0 | >= 2.1 |
| < 1.0.0                   | >= v0.12.0 | >= 1.9 |
---------------------------------------------------

We actually use fluentd 0.14.22 now, but the version of fluent-plugin-grok-parser
is still pinned in 0.3.1. So we need to upgrade the fluent-plugin-grok-parser to
2.1.4.

[1] https://github.com/fluent/fluent-plugin-rewrite-tag-filter
[2] http://paste.openstack.org/show/626145/
[3] https://github.com/fluent/fluent-plugin-grok-parser/issues/32

backport: pike

Change-Id: I47985113fe732569d640a262ca832c6edc8c2bb6
Partial-Bug: #1730664