Install python-openstackclient in kolla-toolbox container
in the past, shade depends on python-ironicclient and
python-ironicclient depends on python-openstackclient. But in the
recent released python-ironiclient 2.0, python-openstackclient is
removed[0]. So python-openstackclient is missing. This will affect all
branches.
Here is a potential issue. we need use upper-constraints.txt
file in kolla-toolbox. install the latest python packages is not a good
idea. Implement this may take much time. Let us fix the gate ASAP and
improve this later.

[0] https://review.openstack.org/#/c/297730/

Partial-Bug: #1732180
Change-Id: I55810b641e388e26129b7fa77f97a8736eb1bb48