Revert "nova-libvirt: fix kvm permission issue"
This is not needed. If we make sure qemu use nova user.
Because nova user in group qemu.

See comment #7 of bug #1715356:
The root cause for this issue is that qemu not run with nova user on
debian.
If we set qemu running with nova user with this patch:
https://review.openstack.org/#/c/525891/
Then we need to revert commit f1b98a4925826333cf22dafe15ef755066c51d48.

This reverts commit f1b98a4925826333cf22dafe15ef755066c51d48.
Closes-bug: #1715356
Depends-on: I36720af0c7d3dd7c69d2404843f54c0991aea1bb

Change-Id: I62fbcf9e4ee5c3170c96576698f4ae8b66db1b74