Suport configuration of Kafka package
This change makes it easier to configure the Kafka package. This
is useful, for example, when installing Monasca which currently
requires an older version of Kafka than that installed here.

This change also fixes an issue with circular symlinks and verifies
the sha512sum of the package. The sha512sum was taken from here:

https://www.apache.org/dist/kafka/1.0.1/kafka_2.12-1.0.1.tgz.sha512

Partially-Implements: blueprint monasca-roles